View2D: Border-Zoom tool
From the Christmas goodie-bag, I've added the border-zoom tool (also known as 'Zoom to Border'). This can be activated using the Shift-ZKEY hotkey (Note: this probably won't be the final hotkey choice, but I'm trying to keep Shift-BKEY free for data-editing tools). 

With this method of zooming, you define a border for the current viewable area to 'fit' into.
* drawing the region with LMB will zoom in the view by focussing in on the defined region 
* drawing the region with RMB will zoom out of the view by shrinking the view into the defined region